requirements.txt
pillow==10.4.0
PyPDF2==3.0.1
